requirements.txt
fastapi==0.115.6
uvicorn==0.33.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.10
alembic==1.12.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
python-multipart
pydantic-settings
pydantic[email]
jinja2
flake8
black